+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+- Add a method to retrieve the list of blogposts related to (ie that
+  share at least one category with) another blogpost.
 - Licences (through their "name" and "content" fields) are now translatable.
 - Report frontend errors through Sentry when a sentry DSN is available in
   Django settings.